# openapi.yaml
openapi: "3.0.3"
info:
  title: CampusXchange API
  version: "1.0.0"
  description: |
    Student peer-to-peer marketplace REST API.

    **Base URL:** `http://localhost:8080`
    **Swagger UI:** `http://localhost:8080/swagger-ui.html`

    ## Authentication
    Protected endpoints require two headers:
    ```
    Authorization: Bearer <access_token>
    X-User-Id: <userId>
    ```
    Both values are returned by `/api/auth/login`, `/api/auth/register`, and `/api/auth/google`.

    ## Real-time Chat (WebSocket)
    Connect via SockJS + STOMP to `http://localhost:8080/ws`.
    - Publish to `/app/chat/send` (include `X-User-Id` as STOMP header)
    - Subscribe to `/topic/chat/{min_uid}_{max_uid}` where the UIDs are sorted ascending
servers:
  - url: http://localhost:8080
    description: Local development
security: []   # overridden per-operation where required
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        college:
          type: string
        studentId:
          type: string
          nullable: true
        profilePhotoUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        rating:
          type: number
          format: double
        totalReviews:
          type: integer
        emailVerified:
          type: boolean
        studentVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
    UserStats:
      type: object
      properties:
        totalListings:
          type: integer
        activeListings:
          type: integer
        soldItems:
          type: integer
        savedItems:
          type: integer
        rating:
          type: number
          format: double
        totalReviews:
          type: integer

    # ── Products ──────────────────────────────────────────────────────────
...
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        originalPrice:
          type: number
          format: decimal
          nullable: true
        category:
          $ref: "#/components/schemas/ProductCategory"
        condition:
          $ref: "#/components/schemas/ProductCondition"
        status:
          $ref: "#/components/schemas/ProductStatus"
        sellerId:
          type: integer
          format: int64
        sellerName:
          type: string
        sellerRating:
          type: number
          format: double
          nullable: true
        college:
          type: string
        location:
          type: string
        imageUrls:
          type: array
          items:
            type: string
            format: uri
        videoUrl:
          type: string
          format: uri
          nullable: true
        modelUrl:
          type: string
          format: uri
          nullable: true
        viewCount:
          type: integer
        rating:
          type: number
          format: double
        totalReviews:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Reviews ───────────────────────────────────────────────────────────
...
    UserRole:
      type: string
      enum: [STUDENT, ADMIN]
    ProductCategory:
      type: string
      enum:
        - ELECTRONICS
        - BOOKS
        - FURNITURE
        - CLOTHING
        - SPORTS
        - STATIONERY
        - FOOD
        - TRANSPORT
        - GADGETS
        - NOTES
        - HOSTEL_ESSENTIALS
        - GAMING
        - OTHER
    ProductCondition:
      type: string
      enum: [NEW, LIKE_NEW, GOOD, FAIR, POOR]
    ProductStatus:
      type: string
      enum: [AVAILABLE, SOLD, RESERVED, DELETED]

    # ── Common ────────────────────────────────────────────────────────────
...
    Page:
      type: object
      description: Spring Data pagination wrapper
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
          description: Current page (0-indexed)
        first:
          type: boolean
        last:
          type: boolean
...
paths:
...
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get a user's public profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "404":
          description: User not found

    put:
      tags: [Users]
      summary: Update own profile
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Unauthenticated
  /api/users/{id}/products:
    get:
      tags: [Users]
      summary: Get all products listed by a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProductDTO"
  /api/users/{id}/reviews:
    get:
      tags: [Users]
      summary: Get all reviews written by a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
  /api/users/{id}/stats:
    get:
      tags: [Users]
      summary: Get activity statistics for a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
  /api/users/search:
    get:
      tags: [Users]
      summary: Search users by full name
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
  /api/users/college/{college}:
    get:
      tags: [Users]
      summary: Get users from a specific college
      parameters:
        - in: path
          name: college
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK

  # ── Products ──────────────────────────────────────────────────────────
...